+++ 2_hello_world/csharp/HelloWorldSubscriber/HelloWorldSubscriber.csproj
