+++ 3_streaming_data/csharp/TemperaturePublisher/TemperaturePublisher.csproj
