+++ 3_streaming_data/csharp/TemperatureSubscriber/TemperatureSubscriber.csproj
