+++ 4_keys_instances/csharp/MonitoringCtrlApplication/MonitoringCtrlApplication.csproj
